Convert Welcome component to TypeScript
Refactored; TypeScript; functional component.
